1306 - now working again with int; so this was a transient
Makes sense; ints and longs are identical on this platform.

+++ 020posix.cc
@@ -61,7 +61,7 @@ COMPILE_PRIM_FUNC(serverc0, primFunc_serverc0, "($port)",
   PERR(setsockopt(sockfd, SOL_SOCKET, SO_REUSEADDR, &dummy, sizeof(dummy)));
   sockaddr_in s;  s.sin_family = AF_INET;   s.sin_addr.s_addr = INADDR_ANY;
   cout << "a: " << toNum(lookup("$port")) << endl;
-  s.sin_port = htons(toNum(lookup("$port")));
+  s.sin_port = htons((int)toNum(lookup("$port")));
   cout << "b: " << s.sin_port << endl;
   PERR(bind(sockfd, (sockaddr*)&s, sizeof(s)));
   PERR(listen(sockfd, 5));
@@ -98,7 +98,7 @@ COMPILE_PRIM_FUNC(serverc1, primFunc_serverc1, "($port)",
   PERR(setsockopt(sockfd, SOL_SOCKET, SO_REUSEADDR, &dummy, sizeof(dummy)));
   sockaddr_in s;  s.sin_family = AF_INET;   s.sin_addr.s_addr = INADDR_ANY;
   cout << "a: " << toNum(lookup("$port")) << endl;
-  s.sin_port = htons(toNum(lookup("$port")));
+  s.sin_port = htons((int)toNum(lookup("$port")));
   cout << "b: " << s.sin_port << endl;
   PERR(bind(sockfd, (sockaddr*)&s, sizeof(s)));
   PERR(listen(sockfd, 5));
